Industrial visit to AIMA Index 2022 Exhibition
AIMA Ambad Industries and Manufacturers' Association is an industrial
body supporting Nashik industrialists urging them to participate in many
industrial activities undertaken for the benefit of industrial community at
large. Over the last 30 years AIMA has strived towards the betterment of
industries, ensuring that all small scale businesses get benefitted from the
same. Today Nashik has transformed itself in to an industrial hub promising
exponential growth in the sector.

In order to give a further boost, AIMA is coming up with its flagship event
AIMA INDEX 2022, an industrial exhibition which is held regularly and is
the 6th edition this year. This exhibition will provide a platform to all
industrialists, displaying the immense potential of the industrial sector in
Nashik.
In the exhibition, there were 300 stalls.
The exhibition was spread into the area of 320000 sq. Mtr.
It was a great place for Business& Networking Opportunities,
Buyer- Sellers Meet, National and
International Delegation from MSMEs and Government Agencies.